==> code/3_classification.Rmd <==
---
title: "3. Classification"
author: "Ross Dahlke"
date: "9/25/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(quanteda)
library(quanteda.textmodels)
```


# Classification

```{r}
coded <- read_csv("~/git/political_donor_motivations/data/posts_coded.csv") %>% 
  gather("topic", "value", -text, na.rm = T) %>% 
  sample_frac(1)

```


```{r}
text_corpus <- corpus(coded$text)
docvars(text_corpus, field = "topic") <- coded$topic
```

```{r}
text_dfm <- dfm(text_corpus, tolower = T, remove = stopwords(), remove_punct = T, remove_symbols = T) %>% 
  dfm_trim(min_termfreqError = 2, min_docfreq = 1) %>% 
  dfm_tfidf()

split_n <- as.integer(nrow(coded) * 2 / 3)
split_n_1 <- split_n + 1

train_data <- coded[1:split_n,]
test_data <- coded[split_n_1:nrow(coded),]

train_dfm <- text_dfm[1:split_n,]
test_dfm <- text_dfm[split_n_1:nrow(coded),]


```

```{r}
nb_classifier <- textmodel_nb(train_dfm, train_data$topic)
```

```{r}
nb_classifier
```

```{r}
predicted <- predict(nb_classifier, test_dfm, type = "class")
```

```{r}
predicted_2 <- cbind(test_data, tibble(predicted))
```

```{r}
predicted_2 %>% 
  mutate(correct = if_else(topic == predicted, 1, 0)) %>% 
  summarize(correct = sum(correct),
            count = n(),
            per = correct / count)
```


```{r}
uncoded <- read_csv("~/git/political_donor_motivations/data/posts_uncoded.csv") %>% 
  mutate(topic = NA)

all_posts <- coded %>% 
  select(-value) %>% 
  rbind(uncoded %>% select(text, topic))

all_corpus <- corpus(all_posts$text)
docvars(all_corpus, field = "topic") <- all_posts$topic
  
all_dfm <- dfm(all_corpus, tolower = T, remove = stopwords(), remove_punct = T, remove_symbols = T) %>% 
  dfm_trim(min_termfreqError = 2, min_docfreq = 1) %>% 
  dfm_tfidf()

train_dfm <- all_dfm[1:split_n,]
train_data <- all_posts[1:split_n,]

uncoded_start <- nrow(all_posts)-nrow(uncoded)+1
uncoded_dfm <- all_dfm[uncoded_start:nrow(all_posts),]
```

```{r}
nb_classifier <- textmodel_nb(train_dfm, train_data$topic)

uncoded_predicted <- cbind(uncoded, tibble(predicted = predict(nb_classifier, uncoded_dfm)))
```

```{r}
total_counts <- uncoded_predicted %>%
  group_by(target) %>%
  summarize(total_count=n())

topic_scores <- uncoded_predicted %>% 
  group_by(target, predicted) %>% 
  summarize(topic_count = n()) %>% 
  left_join(total_counts) %>% 
  mutate(candidate_topic=topic_count/total_count*100, 
         target_topic=topic_count/210*100, 
         topic_score=candidate_topic*target_topic)%>%
  filter(!str_detect(predicted, "other"))
```

```{r}
topic_scores %>% 
  arrange(desc(topic_score))
```


```{r}
saveRDS(topic_scores, "~/git/political_donor_motivations/data/candidate_topic_scores.RDA")
```
