==> code/4_donations.Rmd <==
---
title: "4. Scoring"
author: "Ross Dahlke"
date: "9/25/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
options(scipen = 9999)
```

```{r}
donations <- read_csv("~/git/political_donor_motivations/data/2016_donations.csv") %>%
  filter(type == "Individual") %>% 
  rename(original_source = source) %>% 
  mutate(original_source = tolower(original_source)) %>% 
  filter(str_detect(original_source, c("unitemized|anonymous")) == F) %>% 
  mutate(zip_5 = str_sub(zip, 1, 5),
         refined_source = refinr::n_gram_merge(refinr::key_collision_merge(original_source)),
         source = paste0(refined_source," : ",zip_5)) 

multi_campaign_donors <- donations %>% 
  distinct(source, target) %>% 
  group_by(source) %>% 
  count() %>% 
  filter(n > 1) %>% 
  pull(source)

donations %>% 
  filter(source %in% multi_campaign_donors) %>% 
  write_csv("~/git/political_donor_motivations/data/2016_donations_refined.csv")
```

```{r}
clusters <- read_csv("~/git/political_donor_motivations/data/gephi_output_all.csv") %>% 
  select(source = Id,
         modularity_class)
```

```{r}
donations %>% 
  left_join(clusters) %>% 
  filter(!is.na(modularity_class)) %>% 
  write_csv("~/git/political_donor_motivations/data/2016_donations_clusters_all.csv")
```
