==> code/5_calculations.Rmd <==
---
title: "5. Calculations"
author: "Ross Dahlke"
date: "9/27/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(parallel)
library(doParallel)
library(infer)
options(scipen = 999)
```

```{r}
donations <- read_csv("../data/2016_donations_clusters_all.csv")
uncoded_predicted <- read_csv("~/git/political_donor_motivations/data/bert_uncoded_predictions.csv") %>% 
  rename(predicted = variable)
```

```{r}
n_clusters <- donations %>% 
  # filter(modularity_class <= 12) %>% 
  distinct(modularity_class) %>% 
  count() %>% 
  pull(n)

n_topics <- uncoded_predicted %>% 
  distinct(predicted) %>% 
  count() %>% 
  pull(n)

p_adjustments <- n_clusters * n_topics * 2
```


```{r}
granger_model <- function(cluster,
                          topic,
                          donation_per_threshold = .01,
                          adjustments = p_adjustments){

campaigns <- donations %>% 
  filter(modularity_class == cluster) %>%
  count(target) %>% 
  mutate(per = n / sum(n)) %>% 
  filter(per >= donation_per_threshold) %>% 
  pull(target)

posts_by_topic <- uncoded_predicted %>% 
  filter(target %in% campaigns) %>% 
  group_by(predicted, date) %>% 
  summarize(topic_count = n()) %>% 
  mutate(date = lubridate::mdy(date)) %>% 
  arrange(date)

cluster_donations <- donations %>% 
  filter(modularity_class == cluster) %>% 
  group_by(modularity_class, date) %>% 
  summarize(donation_count = n(),
            donation_amount = sum(contribution)) %>% 
  mutate(date = lubridate::mdy(date)) %>% 
  arrange(date) %>% 
  ungroup()

granger_data <- tibble(date = seq.Date(as.Date("2015-01-01"), as.Date("2016-11-06"), by = 1)) %>% 
  left_join(cluster_donations) %>%  
  left_join(posts_by_topic %>% 
              filter(predicted == topic)) %>% 
  select(date, topic_count, donation_count, donation_amount) %>% 
  mutate_if(is.numeric, replace_na, replace = 0)

best_lag <- tsDyn::lags.select(granger_data %>% select(topic_count, donation_count), lag.max = 8)
order <-  best_lag$BIC_min[2]

if(sum(granger_data$topic_count) == 0 | sum(granger_data$donation_count) == 0){
  
  output <- tibble(cluster = cluster,
       topic = topic,
       result = "no relationship",
       lag = c(order),
       f_statistic = NA,
       p_value = NA)
  
  return(output)
}
  
access_model <- lmtest::grangertest(diff(granger_data$topic_count) ~ diff(granger_data$donation_count), order = order) %>% 
  tibble() %>% 
  filter(row_number() == 2) %>% 
  mutate(model = "access")

consumption_model <- lmtest::grangertest(diff(granger_data$donation_count) ~ diff(granger_data$topic_count), order = order) %>% 
  tibble() %>% 
  filter(row_number() == 2) %>% 
  mutate(model = "consumption")
  
results <- access_model %>% 
  rbind(consumption_model) %>% 
  janitor::clean_names() %>% 
  mutate(modularity_class = cluster,
         topic = topic,
         adjusted_pr_f = p.adjust(pr_f, method = "bonferroni", n = adjustments)) %>% 
  filter(adjusted_pr_f < .05) 

if(nrow(results) == 0){
  result <- "no relationship"
  f_statistic <- NA
  p_value <- NA
}

if(nrow(results) == 1){
  result <- results %>% pull(model)
  f_statistic <- results %>% pull(f)
  p_value <- results %>% pull(pr_f)
}

if(nrow(results) == 2){
  result <- "confounder"
  f_statistic <- NA
  p_value <- NA
}

output <- tibble(cluster = c(cluster),
       topic = c(topic),
       result = c(result),
       lag = c(order),
       f_statistic = c(f_statistic),
       p_value = c(p_value))
  
return(output)

}
```

```{r}
topics <- uncoded_predicted %>% distinct(predicted) %>% pull()
modularity_classes <- donations %>% 
  # filter(modularity_class <= 12) %>%  
  distinct(modularity_class) %>% arrange(modularity_class) %>% pull() 
```

```{r}
n_cores <- detectCores()

registerDoParallel(makeCluster(n_cores - 1))
```

```{r}
granger_results <- foreach(cluster = modularity_classes,
                           .packages = c("dplyr", "tidyr"),
                           .errorhandling = c("pass"),
                           .combine = "rbind") %:%
  foreach(topic = topics,
          .packages = c("dplyr", "tidyr"),
          .errorhandling = c("pass"),
          .combine = "rbind") %dopar%
  granger_model(cluster = cluster,
          topic = topic)

```

```{r}
plot <- granger_results %>% 
  mutate(topic = case_when(
    topic == "voting.liberal" ~ "voting: liberal",
    topic == "voting.conservative" ~ "voting: conservative",
    topic == "raceissues.liberal" ~ "race issues: liberal",
    topic == "prolife" ~ "abortion and women's issues: conservative",
    topic == "prochoice.womenshealth" ~ "abortion and women's issues: liberal",
    topic == "other" ~ "other",
    topic == "marijuana.pro" ~ "cannabis: liberal",
    topic == "lgbt.pro" ~ "lgbtq+ issues: liberal",
    topic == "lgbt.anti" ~ "lgbtq+ issues: conservative",
    topic == "king.veterans" ~ "veterans issues: bipartisan",
    topic == "infrastructure.roads.transportation" ~ "infrastructure: liberal",
    topic == "healthcare.liberal" ~ "healthcare: liberal",
    topic == "healthcare.conservative" ~ "healthcare: conservative",
    topic == "healthcare.bipartisan" ~ "healthcare: bipartisan",
    topic == "guncontrol.liberal" ~ "guns: liberal",
    topic == "guncontrol.conservative" ~ "guns: conservative",
    topic == "gerrymandering.anti" ~ "gerrymandering: liberal",
    topic == "environment.climatechange" ~ "environment: liberal",
    topic == "education.liberal" ~ "education: liberal",
    topic == "education.conservative" ~ "education: conservative",
    topic == "economic.liberal" ~ "economy: liberal",
    topic == "economic.conservative" ~ "economy: conservative",
    topic == "drugabuse" ~ "drug abuse: bipartisan",
    topic == "domestic.violence" ~ "domestic violence: bipartisan",
    topic == "criminaljusticereform" ~ "criminal justice reform: liberal",
    topic == "crime.police" ~ "crime and police: conservative",
    topic == "campaignfinance.corruption.opengovernment" ~ "governance: liberal"
  )) %>% 
  mutate(result = factor(result, c("access", "consumption", "confounder", "no relationship"))) %>% 
  ggplot(aes(cluster, reorder(topic, desc(topic)), fill = result)) +
  geom_raster() +
  scale_x_continuous(breaks = seq(0, 12, 1), position = "top") +
  scale_fill_grey() +
  labs(x = "donor coalition",
       y = "topic",
       fill = "donor motivation model",
       caption = "p-values adjusted with Bonferroni's method") +
  theme_bw() +
  theme(legend.position = "bottom") +
  guides(fill = guide_legend(nrow = 1))
```

```{r}
ggsave("../tables_and_figures/fig_2.jpg", plot, height = 6.9, width = 7.3)
```

```{r}
granger_results %>% 
  mutate(topic = case_when(
    topic == "voting.liberal" ~ "voting: liberal",
    topic == "voting.conservative" ~ "voting: conservative",
    topic == "raceissues.liberal" ~ "race issues: liberal",
    topic == "prolife" ~ "abortion and women's issues: conservative",
    topic == "prochoice.womenshealth" ~ "abortion and women's issues: liberal",
    topic == "other" ~ "other",
    topic == "marijuana.pro" ~ "cannabis: liberal",
    topic == "lgbt.pro" ~ "lgbtq+ issues: liberal",
    topic == "lgbt.anti" ~ "lgbtq+ issues: conservative",
    topic == "king.veterans" ~ "veterans issues: bipartisan",
    topic == "infrastructure.roads.transportation" ~ "infrastructure: liberal",
    topic == "healthcare.liberal" ~ "healthcare: liberal",
    topic == "healthcare.conservative" ~ "healthcare: conservative",
    topic == "healthcare.bipartisan" ~ "healthcare: bipartisan",
    topic == "guncontrol.liberal" ~ "guns: liberal",
    topic == "guncontrol.conservative" ~ "guns: conservative",
    topic == "gerrymandering.anti" ~ "gerrymandering: liberal",
    topic == "environment.climatechange" ~ "environment: liberal",
    topic == "education.liberal" ~ "education: liberal",
    topic == "education.conservative" ~ "education: conservative",
    topic == "economic.liberal" ~ "economy: liberal",
    topic == "economic.conservative" ~ "economy: conservative",
    topic == "drugabuse" ~ "drug abuse: bipartisan",
    topic == "domestic.violence" ~ "domestic violence: bipartisan",
    topic == "criminaljusticereform" ~ "criminal justice reform: liberal",
    topic == "crime.police" ~ "crime and police: conservative",
    topic == "campaignfinance.corruption.opengovernment" ~ "governance: liberal"
  )) %>%
  write_csv("../data/granger_results.csv")
```

# Contributor size

```{r}
cluster_category <- granger_results %>% 
  group_by(cluster) %>% 
  summarize(category = case_when(
    any(result == "access") & any(result == "consumption") ~ "access and consumption",
    any(result == "access") ~ "access",
    any(result == "consumption") ~ "consumption",
    any(result == "confounder") ~ "confounder",
    TRUE ~ "0_no_relationship"
  ))
```


```{r}
donor_totals <- donations %>% 
  group_by(source, modularity_class) %>% 
  summarize(total_contributed = sum(contribution, na.rm = T),
            n_contributed = n()) %>% 
  left_join(cluster_category, by = c("modularity_class" = "cluster")) %>% 
  filter(n_contributed > 1) %>% 
  arrange(desc(total_contributed))
```

```{r}
access_bootstrap <- donor_totals %>% 
  mutate(category = if_else(category != "access", "not access", "access")) %>% 
  specify(total_contributed ~ category) %>% 
    generate(reps = 1000, type = "bootstrap") %>% 
    calculate(stat = "diff in means", order = c("access", "not access")) 

access_output <- access_bootstrap %>% 
  get_ci() %>% 
  cbind(access_bootstrap %>% 
          summarize(mean_diff = mean(stat))) %>% 
  cbind(access_bootstrap %>% 
          get_p_value(obs_stat = 0, direction = "two_sided")) %>% 
  mutate(model = "access")

```

```{r}
consumption_bootstrap <- donor_totals %>% 
  mutate(category = if_else(category != "consumption", "not consumption", "consumption")) %>% 
  specify(total_contributed ~ category) %>% 
    generate(reps = 1000, type = "bootstrap") %>% 
    calculate(stat = "diff in means", order = c("consumption", "not consumption")) 

consumption_output <- consumption_bootstrap %>% 
  get_ci() %>% 
  cbind(consumption_bootstrap %>% 
          summarize(mean_diff = mean(stat))) %>% 
  cbind(consumption_bootstrap %>% 
          get_p_value(obs_stat = 0, direction = "two_sided")) %>% 
  mutate(model = "consumption")
```

```{r}
access_output %>% 
  rbind(consumption_output) %>% 
  write_csv("../data/contribution_size_permutation.csv")
```


```{r}
lm(total_contributed ~ category, data = donor_totals) %>% 
  summary()
```

```{r}
lm(n_contributed ~ category, data = donor_totals) %>% 
  summary()
```

```{r}
access_calcs_clusters <- donations %>% 
  group_by(source, modularity_class) %>% 
  summarize(total_contributed = sum(contribution, na.rm = T),
            n_contributed = n()) %>% 
  left_join(cluster_category, by = c("modularity_class" = "cluster")) %>% 
  mutate(category = if_else(category == "access and consumption", "access", category)) %>% 
  group_by(modularity_class, category) %>% 
  summarize(n_donors = n(),
            n_contributions = sum(n_contributed),
            dollars_contributed = sum(total_contributed)) %>% 
  ungroup() %>% 
  mutate(n_donors_percent = n_donors / sum(n_donors),
         n_contributions_percent = n_contributions / sum(n_contributions),
         dollars_contributed_percent = dollars_contributed / sum(dollars_contributed))

access_calcs_total <- donations %>% 
  group_by(source, modularity_class) %>% 
  summarize(total_contributed = sum(contribution, na.rm = T),
            n_contributed = n()) %>% 
  left_join(cluster_category, by = c("modularity_class" = "cluster")) %>% 
  mutate(category = if_else(category == "access and consumption", "access", category)) %>% 
  group_by(category) %>% 
  summarize(n_donors = n(),
            n_contributions = sum(n_contributed),
            dollars_contributed = sum(total_contributed)) %>% 
  ungroup() %>% 
  mutate(n_donors_percent = n_donors / sum(n_donors),
         n_contributions_percent = n_contributions / sum(n_contributions),
         dollars_contributed_percent = dollars_contributed / sum(dollars_contributed))
```

```{r}
consumption_calcs_clusters <- donations %>% 
  group_by(source, modularity_class) %>% 
  summarize(total_contributed = sum(contribution, na.rm = T),
            n_contributed = n()) %>% 
  left_join(cluster_category, by = c("modularity_class" = "cluster")) %>% 
  mutate(category = if_else(category == "access and consumption", "consumption", category)) %>% 
  group_by(modularity_class, category) %>% 
  summarize(n_donors = n(),
            n_contributions = sum(n_contributed),
            dollars_contributed = sum(total_contributed)) %>% 
  ungroup() %>% 
  mutate(n_donors_percent = n_donors / sum(n_donors),
         n_contributions_percent = n_contributions / sum(n_contributions),
         dollars_contributed_percent = dollars_contributed / sum(dollars_contributed))

access_calcs_total <- donations %>% 
  group_by(source, modularity_class) %>% 
  summarize(total_contributed = sum(contribution, na.rm = T),
            n_contributed = n()) %>% 
  left_join(cluster_category, by = c("modularity_class" = "cluster")) %>% 
  mutate(category = if_else(category == "access and consumption", "consumption", category)) %>% 
  group_by(category) %>% 
  summarize(n_donors = n(),
            n_contributions = sum(n_contributed),
            dollars_contributed = sum(total_contributed)) %>% 
  ungroup() %>% 
  mutate(n_donors_percent = n_donors / sum(n_donors),
         n_contributions_percent = n_contributions / sum(n_contributions),
         dollars_contributed_percent = dollars_contributed / sum(dollars_contributed))
```

```{r}
gexf <- rgexf::read.gexf("../data/gephi_output_all.gexf") 

gephi_nodes <- gexf$nodes %>% 
  cbind(gexf$nodesVizAtt$position) %>% 
  left_join(read_csv("~/git/political_donor_motivations/data/gephi_output_all.csv") %>% rename(id = Id)) %>% 
  left_join(cluster_category, by = c("modularity_class" = "cluster")) %>% 
  mutate(x = x - .5,
         x = scales::rescale(x, to = c(-1, 1)),
         x_abs = abs(x)) 
```

```{r}
access_bootstrap <- gephi_nodes %>% 
  mutate(category = if_else(category != "access", "not access", "access")) %>% 
  specify(x_abs ~ category) %>% 
    generate(reps = 1000, type = "bootstrap") %>% 
    calculate(stat = "diff in means", order = c("access", "not access")) 

access_output <- access_bootstrap %>% 
  get_ci() %>% 
  cbind(access_bootstrap %>% 
          summarize(mean_diff = mean(stat))) %>% 
  cbind(access_bootstrap %>% 
          get_p_value(obs_stat = 0, direction = "two_sided")) %>% 
  mutate(model = "access")

```

```{r}
consumption_bootstrap <- gephi_nodes %>% 
  mutate(category = if_else(category != "consumption", "not consumption", "consumption")) %>% 
  specify(x_abs ~ category) %>% 
    generate(reps = 1000, type = "bootstrap") %>% 
    calculate(stat = "diff in means", order = c("consumption", "not consumption")) 

consumption_output <- consumption_bootstrap %>% 
  get_ci() %>% 
  cbind(consumption_bootstrap %>% 
          summarize(mean_diff = mean(stat))) %>% 
  cbind(consumption_bootstrap %>% 
          get_p_value(obs_stat = 0, direction = "two_sided")) %>% 
  mutate(model = "consumption")
```

```{r}
access_output %>% 
  rbind(consumption_output) %>% 
  write_csv("../data/contribution_polarization_permutation.csv")
```


```{r}
lm(x_abs ~ category, data = gephi_nodes) %>% 
  summary()
```

# Graphs

```{r}
donor_size_histogram <- gephi_nodes %>% 
  filter(category == "consumption") %>% 
  mutate(category = "Consumption-Oriented") %>% 
  rbind(gephi_nodes %>% 
    filter(category == "access") %>% 
    mutate(category = "Access-Oriented")) %>% 
  rbind(gephi_nodes %>% 
    filter(str_detect("0_no_relationship", category)) %>% 
    mutate(category = "No Relationship")) %>% 
  left_join(donor_totals %>% select(-category), by = c("id" = "source")) %>% 
  group_by(category) %>% 
  mutate(mean = mean(total_contributed, na.rm = T),
         low_quantile = quantile(total_contributed, .025, na.rm = T),
         high_quantile = quantile(total_contributed, .975, na.rm = T)) %>% 
  ggplot(aes(total_contributed, fill = category)) +
  geom_histogram(bins = 25) +
  geom_vline(aes(xintercept = mean), linetype = "dashed") +
  # geom_rect(aes(xmin = low_quantile, xmax = high_quantile, ymin = -Inf, ymax = .1), alpha = .2, fill = "darkblue") +
  scale_x_log10(labels = scales::dollar_format()) +
  scale_fill_grey() +
  facet_wrap(.~category, ncol = 1, scale = "free_y") +
  labs(title = "Distribution of Total Contribution Amounts by Motivational Model",
       subtitle = "dashed line = mean for each group",
       x = "total dollars contributed",
       y = "count",
       fill = "donor motivational model",
       caption = "x-axis is on a log scale") +
  theme_bw()


```

```{r}
ggsave("../tables_and_figures/fig_donor_sizes.jpg", donor_size_histogram, height = 6.9, width = 7.3)
```

```{r}
node_position <- gephi_nodes %>% 
  filter(category == "consumption") %>% 
  mutate(category = "Consumption-Oriented") %>% 
  rbind(gephi_nodes %>% 
    filter(category == "access") %>% 
    mutate(category = "Access-Oriented")) %>% 
  rbind(gephi_nodes %>% 
    filter(str_detect("0_no_relationship", category)) %>% 
    mutate(category = "No Relationship")) %>% 
  ggplot(aes(x, fill = category)) +
    geom_histogram(bins = 25) +
    scale_fill_grey() +
    facet_wrap(.~category, ncol = 1, scale = "free_y") +
    labs(title = "Distribution of Ideological Positions of Donors",
         subtitle = "-1 = most liberal; 1 = most conservative",
         x = "Ideological Position from Liberal to Conservative",
         y = "count",
         fill = "donor motivational model") +
    theme_bw()
```

```{r}
ggsave("../tables_and_figures/fig_node_position.jpg", node_position, height = 6.9, width = 7.3)
```

```{r}
node_position_absolute <- gephi_nodes %>% 
  filter(category == "consumption") %>% 
  mutate(category = "Consumption-Oriented") %>% 
  rbind(gephi_nodes %>% 
    filter(category == "access") %>% 
    mutate(category = "Access-Oriented")) %>% 
  rbind(gephi_nodes %>% 
    filter(str_detect("0_no_relationship", category)) %>% 
    mutate(category = "No Relationship")) %>% 
  group_by(category) %>% 
  mutate(mean = mean(x_abs, na.rm = T),
         low_quantile = quantile(x_abs, .025, na.rm = T),
         high_quantile = quantile(x_abs, .975, na.rm = T)) %>% 
  ggplot(aes(x_abs, fill = category)) +
    geom_histogram(bins = 25) +
    geom_vline(aes(xintercept = mean), linetype = "dashed") +
    # geom_rect(aes(xmin = low_quantile, xmax = high_quantile, ymin = -Inf, ymax = .1), alpha = .2, fill = "darkblue") +
    scale_fill_grey() +
    facet_wrap(.~category, ncol = 1, scale = "free_y") +
    labs(title = "Distribution of the Absolute Ideological Positions of Donors",
         subtitle = "0 = least polarized; 1 = most polarized",
         x = "Absolute Value of Ideological Position",
         y = "count",
         fill = "donor motivational model") +
    theme_bw()
```

```{r}
ggsave("../tables_and_figures/fig_node_position_absolute.jpg", node_position_absolute, height = 6.9, width = 7.3)
```

```{r}
donations %>% 
  filter(party != "Ind") %>% 
  left_join(gephi_nodes %>% select(id, category), by = c("source" = "id")) %>% 
  group_by(source, category, party) %>% 
    summarize(total_donated = sum(contribution)) %>% 
  pivot_wider(id_cols = c("source", "category"), names_from = "party", values_from = "total_donated") %>% 
  mutate(Rep = replace_na(Rep, 0),
         Dem = replace_na(Dem, 0),
         per_rep = Rep / sum(Rep + Dem),
         rep = if_else(per_rep > .5, 1, 0),
         dem = if_else(per_rep < .5, 1, 0)) %>% 
  group_by(category) %>% 
  summarize(rep = sum(rep),
            dem = sum(dem)) %>% 
  mutate(n = rep + dem,
         per_rep = rep / n,
         per_dem = dem / n) %>% 
  filter(category != "access and consumption") %>% 
  write_csv("../tables_and_figures/party_category_breakdown.csv")
```

# Example time series

Access-oriented

```{r}
access_sim <- tibble(day = seq(1, 30, 1),
       donations = rnorm(30, 1, .1),
       posts = rnorm(30, 1, .1)) %>% 
  mutate(donations = if_else(row_number() %% 7 == 0, donations + 1, donations),
         posts = if_else(lag(donations, n = 2)  > 1.4, posts + 1, posts),
         posts = replace_na(posts, 1)) %>% 
  pivot_longer(cols = c("donations", "posts")) %>% 
  ggplot(aes(day, value)) +
  geom_line() +
  scale_x_continuous(breaks = seq(1, 30, 2)) +
  facet_wrap(.~name, ncol = 1) + 
  labs(title = "Simulated Access-Oriented Model",
       subtitle = "A spike in donations predicts a spike in posts 2 days later",
       x = "Day",
       y = "Value") +
  theme_bw() + 
  theme(axis.text.y=element_blank(),
        axis.ticks.y=element_blank())

ggsave("../tables_and_figures/simulated_access_time_series.jpg", access_sim, height = 4, width = 7.3)
```

Consumption Oriented 

```{r}
consumption_sim <- tibble(day = seq(1, 30, 1),
       donations = rnorm(30, 1, .1),
       posts = rnorm(30, 1, .1)) %>% 
  mutate(posts = if_else(row_number() %% 7 == 0, posts + 1, posts),
         donations = if_else(lag(posts, n = 1)  > 1.4, donations + 1, donations),
         donations = replace_na(donations, 1)) %>% 
  pivot_longer(cols = c("donations", "posts")) %>% 
  ggplot(aes(day, value)) +
  geom_line() +
  scale_x_continuous(breaks = seq(1, 30, 2)) +
  facet_wrap(.~name, ncol = 1) + 
  labs(title = "Simulated Consumption-Oriented Model",
       subtitle = "A spike in posts predicts a spike in donations 1 day later",
       x = "Day",
       y = "Value") +
  theme_bw() + 
  theme(axis.text.y=element_blank(),
        axis.ticks.y=element_blank())

ggsave("../tables_and_figures/simulated_consumption_time_series.jpg", consumption_sim, height = 4, width = 7.3)
```
